+++ packages/grid/data-grid/src/DataGrid.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 // @ts-expect-error need to migrate helpers to TypeScript
-import { createClientRender, ErrorBoundary } from 'test/utils';
+import { createClientRender, fireEvent, screen, ErrorBoundary } from 'test/utils';
 import { useFakeTimers } from 'sinon';
 import { expect } from 'chai';
 import { DataGrid } from '@material-ui/data-grid';
@@ -264,6 +264,66 @@ describe('<DataGrid />', () => {
         expect(getColumnValues()).to.deep.equal(['Puma', 'Nike', 'Adidas']);
         setProps({ direction: 'asc', sortedRows: [1, 0, 2] });
         expect(getColumnValues()).to.deep.equal(['Puma', 'Nike', 'Adidas'].reverse());
+      });
+    });
+
+    describe('sorting', () => {
+      it('should sort when clicking the header cell', () => {
+        render(
+          <div style={{ width: 300, height: 300 }}>
+            <DataGrid {...defaultProps} />
+          </div>,
+        );
+        const header = screen.getByRole('columnheader', { name: 'brand' });
+        expect(getColumnValues()).to.deep.equal(['Nike', 'Adidas', 'Puma']);
+        fireEvent.click(header);
+        expect(getColumnValues()).to.deep.equal(['Adidas', 'Nike', 'Puma']);
+        fireEvent.click(header);
+        expect(getColumnValues()).to.deep.equal(['Puma', 'Nike', 'Adidas']);
+      });
+
+      it('should keep rows sorted when rows prop change', () => {
+        interface TestCaseProps {
+          rows: any[];
+        }
+        const TestCase = (props: TestCaseProps) => {
+          const { rows } = props;
+          return (
+            <div style={{ width: 300, height: 300 }}>
+              <DataGrid
+                {...defaultProps}
+                rows={rows}
+                sortModel={[
+                  {
+                    field: 'brand',
+                    sort: 'asc',
+                  },
+                ]}
+              />
+            </div>
+          );
+        };
+
+        const { setProps } = render(<TestCase rows={defaultProps.rows} />);
+        expect(getColumnValues()).to.deep.equal(['Adidas', 'Nike', 'Puma']);
+
+        setProps({
+          rows: [
+            {
+              id: 3,
+              brand: 'Asics',
+            },
+            {
+              id: 4,
+              brand: 'RedBull',
+            },
+            {
+              id: 5,
+              brand: 'Hugo',
+            },
+          ],
+        });
+        expect(getColumnValues()).to.deep.equal(['Asics', 'Hugo', 'RedBull']);
       });
     });
   });
